requirements.txt
asgiref==3.5.0
Django==4.0.4
django-cors-headers==3.11.0
django-js-asset==2.0.0
django-mptt==0.13.4
djangorestframework==3.13.1
drf-writable-nested==0.6.3
flake8==4.0.1
mccabe==0.6.1
psycopg2==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2022.1
sqlparse==0.4.2
